Hotels: Fix relay/unused-fields in HotelDetailScreen
Refactor countBookingPrice function into HotelDetailScreen since it is
not used elsewhere anymore
